Wait for pushState to complete on AJAX partial page load
The events_listings_day steps fail on Travis as WebDriver does not
complete the AJAX page load before the URL is asserted. Switch to new
"And I wait until I am on the listings page for 22/08/2011" with
a timeout to verify that the pushState completes, rather than asserting
the URL straight away.