As an administrator I can configure the festival dates
Support configurable festival dates, with a new settings page
that allows authenticated users to set dates for the next
festival. Dates are stored with the rails-settings-cached gem
and used to render the calendar and related views.

Also added the forms and alerts CSS from twitter bootstrap to
style settings pages - assign @use_bootstrap=true to include
for other pages.